Rails 5 upgrade: remove clean_empty_strings.rb
This removes `clean_empty_strings.rb` whose purpose seems to be to
replace empty strings with nils on save (or specifically, on type cast
for DB writes).

For now I'm just going to remove it, but if we care about the behavior
we should replicate it using something more standard like
https://github.com/rmm5t/strip_attributes.

In the meantime, I added a spec on master that exercises this behavior,
so it'll remind us to revisit this.