Rails 5 upgrade: replace cancan with cancancan v2
The cancan gem is unmaintained and has been superseded by cancancan
which has support for Rails 6. I'm locking the version to ~> 2 because
while 3 adds support for Rails 6, it contains breaking API changes
which I'd like to defer making for now, due to:

The proximate cause of updating this right now is because I'm seeing
some very odd behavior where if helper_method is used in a controller
to declare a helper method **after** a call to the cancan method
`load_and_authorize_resource`, `helper_method` is never actually
called. But if the `helper_method` call is moved to before the cancan
call, it does work.

I have no idea why this would be, but this gem needs to be upgraded
anyway so this seems like a good first troubleshooting step.