Unsqueel: replace (almost) all uses of `where`
This is a big one. It replaces most Squeel `.where{}` calls with
vanilla AR where calls.

This was a largely mechanical operation and so I don't have 100%
confidence that the resulting code is fully correct. In particular, I
suspect there are places where the predicate of the where clause was
something special to Squeel but AR will not know what to do with. An
example from `incidents_list_helper.rb`:

The original Squeel call was:

`(incident_id == incidents_incidents.id)`

…which I mechanically converted to:

`incident_id: incidents_incidents.id`

…however I do not think `incidents_incidents.id` is a valid method call
in the context of that file. And there are a few other similar
instances where the predicates were more complex than a simple value or
booleans/nil.

Unfortunately time constraints are such that I couldn't puzzle out the
correct equivalent AR-based queries to replace these cases, so I left
them in place where they'll (hopefully) cause a visible failure at the
time of use, and can be given more attention and remedied at that point.

In addition to those unknown cases, there are a couple others (which
are noted with `NOTE:`s) where I suspect a bug or otherwise incorrect
usage in the original Squeel calls.

In these cases, rather than taking a potentially-bad guess at the
original intent, I've inserted an exception raise. This will force any
future callers of that code to take a closer look at them and figure
out the true intended functionality.