SerializedColumns Migration: Step 1
The goal is to replace the use of Postgres `hstore` columns with
`jsonb` columns, which Rails has native type casting support for. Then
we'll be able to remove the custom type casting code from
SerializedColumns which has tight coupling into AR internals.

This commit is step 1: create `jsonb` columns and copy values for each
`serialized_accessor` on each model and record that use them.

This migration calls `save!` on every record, so if there is invalid
data in the database somehow, it will fail and roll back so that can be
corrected (or worked around). The entire migration is wrapped in a
transaction to make this whole operation atomic.

I've tested this by running it on the Roster::Region record created by
`rake db:seed`. Results (note the string numerical values on the hstore
version versus the numerical values on the vanilla jsonb column):

    >> r.config
    => {"scheduler_flex_day_start"=>"28800", "scheduler_flex_night_start"=>"72000"}
    >> r.config_jsonb
    => {"scheduler_flex_day_start"=>28800, "scheduler_flex_night_start"=>72000}